requirements.txt
langchain
langchain-community
pypdf
langchain-text-splitters
langchain
lxml
langchain-openai
python-dotenv
chromadb
langchain-ollama
sentence-transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain_groq 
langserve
fastapi
uvicorn
sse_starlette